Split out event specs
