+++ test/watcher.js
@@ -19,437 +19,6 @@ describe('watcher', function () {
     await cleanupFixtureDir(fixtureDir)
   })
 
-  describe('watching a directory', function () {
-    beforeEach(async function () {
-      await watcher.configure({mainLog: mainLogFile, workerLog: workerLogFile})
-    })
-
-    describe('events', function () {
-      let errors, events, sub
-
-      beforeEach(async function () {
-        errors = []
-        events = []
-
-        sub = await watcher.watch(watchDir, (err, es) => {
-          errors.push(err)
-          events.push(...es)
-        })
-        subs.push(sub)
-      })
-
-      function specMatches (spec, event) {
-        return (spec.type === undefined || event.type === spec.type) &&
-          (spec.kind === undefined || event.kind === spec.kind) &&
-          (event.oldPath === undefined || event.oldPath === spec.oldPath) &&
-          (event.newPath === (spec.newPath || ''))
-      }
-
-      function eventMatching (spec) {
-        const isMatch = specMatches.bind(null, spec)
-        return function () {
-          return events.some(isMatch)
-        }
-      }
-
-      function allEventsMatching (...specs) {
-        const remaining = new Set(specs)
-
-        return function () {
-          for (const event of events) {
-            for (const spec of remaining) {
-              if (specMatches(spec, event)) {
-                remaining.delete(spec)
-              }
-            }
-          }
-
-          return remaining.size === 0
-        }
-      }
-
-      function orderedEventsMatching (...specs) {
-        return function () {
-          let specIndex = 0
-
-          for (const event of events) {
-            if (specs[specIndex] && specMatches(specs[specIndex], event)) {
-              specIndex++
-            }
-          }
-
-          return specIndex >= specs.length
-        }
-      }
-
-      it('when a file is created', async function () {
-        const createdFile = path.join(watchDir, 'file.txt')
-        await fs.writeFile(createdFile, 'contents')
-
-        await until('the creation event arrives', eventMatching({
-          type: 'created',
-          kind: 'file',
-          oldPath: createdFile
-        }))
-      })
-
-      it('when a file is modified', async function () {
-        const modifiedFile = path.join(watchDir, 'file.txt')
-        await fs.writeFile(modifiedFile, 'initial contents\n')
-
-        await until('the creation event arrives', eventMatching({
-          type: 'created',
-          kind: 'file',
-          oldPath: modifiedFile
-        }))
-
-        await fs.appendFile(modifiedFile, 'changed contents\n')
-        await until('the modification event arrives', eventMatching({
-          type: 'modified',
-          kind: 'file',
-          oldPath: modifiedFile
-        }))
-      })
-
-      it('when a file is renamed', async function () {
-        const oldPath = path.join(watchDir, 'old-file.txt')
-        await fs.writeFile(oldPath, 'initial contents\n')
-
-        await until('the creation event arrives', eventMatching({
-          type: 'created',
-          kind: 'file',
-          oldPath,
-          newPath: ''
-        }))
-
-        const newPath = path.join(watchDir, 'new-file.txt')
-
-        await fs.rename(oldPath, newPath)
-
-        await until('the rename event arrives', eventMatching({
-          type: 'renamed',
-          kind: 'file',
-          oldPath,
-          newPath
-        }))
-      })
-
-      it('when a file is renamed from outside of the watch root in', async function () {
-        const outsideFile = path.join(fixtureDir, 'file.txt')
-        const insideFile = path.join(watchDir, 'file.txt')
-
-        await fs.writeFile(outsideFile, 'contents')
-        await fs.rename(outsideFile, insideFile)
-
-        await until('the creation event arrives', eventMatching({
-          type: 'created',
-          kind: 'file',
-          oldPath: insideFile
-        }))
-      })
-
-      it('when a file is renamed from inside of the watch root out ^windows', async function () {
-        const outsideFile = path.join(fixtureDir, 'file.txt')
-        const insideFile = path.join(watchDir, 'file.txt')
-        const flagFile = path.join(watchDir, 'flag.txt')
-
-        await fs.writeFile(insideFile, 'contents')
-
-        await until('the creation event arrives', eventMatching({
-          type: 'created',
-          kind: 'file',
-          oldPath: insideFile
-        }))
-
-        await fs.rename(insideFile, outsideFile)
-        await fs.writeFile(flagFile, 'flag 1')
-
-        await until('the flag file event arrives', eventMatching({
-          type: 'created',
-          kind: 'file',
-          oldPath: flagFile
-        }))
-
-        // Trigger another batch of events on Linux
-        await fs.writeFile(flagFile, 'flag 2')
-
-        await until('the deletion event arrives', eventMatching({
-          type: 'deleted',
-          kind: 'file',
-          oldPath: insideFile
-        }))
-      })
-
-      it('when a file is deleted', async function () {
-        const deletedPath = path.join(watchDir, 'file.txt')
-        await fs.writeFile(deletedPath, 'initial contents\n')
-
-        await until('the creation event arrives', eventMatching({
-          type: 'created',
-          kind: 'file',
-          oldPath: deletedPath
-        }))
-
-        await fs.unlink(deletedPath)
-
-        await until('the deletion event arrives', eventMatching({
-          type: 'deleted',
-          oldPath: deletedPath
-        }))
-      })
-
-      it('understands coalesced creation and deletion events', async function () {
-        const deletedPath = path.join(watchDir, 'deleted.txt')
-        const recreatedPath = path.join(watchDir, 'recreated.txt')
-        const createdPath = path.join(watchDir, 'created.txt')
-
-        await fs.writeFile(deletedPath, 'initial contents\n')
-        await until('file creation event arrives', eventMatching(
-          {type: 'created', kind: 'file', oldPath: deletedPath}
-        ))
-
-        await fs.unlink(deletedPath)
-        await fs.writeFile(recreatedPath, 'initial contents\n')
-        await fs.unlink(recreatedPath)
-        await fs.writeFile(recreatedPath, 'newly created\n')
-        await fs.writeFile(createdPath, 'and another\n')
-
-        await until('all events arrive', orderedEventsMatching(
-          {type: 'deleted', oldPath: deletedPath},
-          {type: 'created', kind: 'file', oldPath: recreatedPath},
-          {type: 'deleted', kind: 'file', oldPath: recreatedPath},
-          {type: 'created', kind: 'file', oldPath: recreatedPath},
-          {type: 'created', kind: 'file', oldPath: createdPath}
-        ))
-      })
-
-      it('correlates rapid file rename events', async function () {
-        const oldPath0 = path.join(watchDir, 'old-file-0.txt')
-        const oldPath1 = path.join(watchDir, 'old-file-1.txt')
-        const oldPath2 = path.join(watchDir, 'old-file-2.txt')
-        const newPath0 = path.join(watchDir, 'new-file-0.txt')
-        const newPath1 = path.join(watchDir, 'new-file-1.txt')
-        const newPath2 = path.join(watchDir, 'new-file-2.txt')
-
-        await Promise.all(
-          [oldPath0, oldPath1, oldPath2].map(oldPath => fs.writeFile(oldPath, 'original\n'))
-        )
-        await until('all creation events arrive', allEventsMatching(
-          {type: 'created', kind: 'file', oldPath: oldPath0},
-          {type: 'created', kind: 'file', oldPath: oldPath1},
-          {type: 'created', kind: 'file', oldPath: oldPath2}
-        ))
-
-        await Promise.all([
-          fs.rename(oldPath0, newPath0),
-          fs.rename(oldPath1, newPath1),
-          fs.rename(oldPath2, newPath2)
-        ])
-
-        await until('all rename events arrive', allEventsMatching(
-          {type: 'renamed', kind: 'file', oldPath: oldPath0, newPath: newPath0},
-          {type: 'renamed', kind: 'file', oldPath: oldPath1, newPath: newPath1},
-          {type: 'renamed', kind: 'file', oldPath: oldPath2, newPath: newPath2}
-        ))
-      })
-
-      it('when a directory is created', async function () {
-        const subdir = path.join(watchDir, 'subdir')
-        await fs.mkdirs(subdir)
-
-        await until('directory creation event arrives', eventMatching(
-          {type: 'created', kind: 'directory', oldPath: subdir}
-        ))
-      })
-
-      it('when a directory is renamed', async function () {
-        const oldDir = path.join(watchDir, 'subdir')
-        const newDir = path.join(watchDir, 'newdir')
-
-        await fs.mkdirs(oldDir)
-        await until('directory creation event arrives', eventMatching(
-          {type: 'created', kind: 'directory', oldPath: oldDir}
-        ))
-
-        await fs.rename(oldDir, newDir)
-        await until('directory rename event arrives', eventMatching(
-          {type: 'renamed', kind: 'directory', oldPath: oldDir, newPath: newDir}
-        ))
-      })
-
-      it('when a directory is deleted', async function () {
-        const subdir = path.join(watchDir, 'subdir')
-        await fs.mkdirs(subdir)
-        await until('directory creation event arrives', eventMatching(
-          {type: 'created', kind: 'directory', oldPath: subdir}
-        ))
-
-        await fs.rmdir(subdir)
-        await until('directory deletion event arrives', eventMatching(
-          {type: 'deleted', oldPath: subdir}
-        ))
-      })
-
-      it('when a directory is deleted and a file is created in its place', async function () {
-        const reusedPath = path.join(watchDir, 'reused')
-        await fs.mkdir(reusedPath)
-        await until('directory creation event arrives', eventMatching(
-          {type: 'created', kind: 'directory', oldPath: reusedPath}
-        ))
-
-        await fs.rmdir(reusedPath)
-        await fs.writeFile(reusedPath, 'IMMA FILE NOW, SURPRIIIISE\n')
-
-        await until('deletion and creation events arrive', orderedEventsMatching(
-          {type: 'deleted', oldPath: reusedPath},
-          {type: 'created', kind: 'file', oldPath: reusedPath}
-        ))
-      })
-
-      it('when a directory is deleted and a file is renamed in its place', async function () {
-        const reusedPath = path.join(watchDir, 'reused')
-        const oldFilePath = path.join(watchDir, 'oldfile')
-
-        await Promise.all([
-          fs.mkdirs(reusedPath),
-          fs.writeFile(oldFilePath, 'original\n')
-        ])
-        await until('directory and file creation events arrive', allEventsMatching(
-          {type: 'created', kind: 'directory', oldPath: reusedPath},
-          {type: 'created', kind: 'file', oldPath: oldFilePath}
-        ))
-
-        await fs.rmdir(reusedPath)
-        await fs.rename(oldFilePath, reusedPath)
-
-        await until('deletion and rename events arrive', allEventsMatching(
-          {type: 'deleted', oldPath: reusedPath},
-          {type: 'renamed', kind: 'file', oldPath: oldFilePath, newPath: reusedPath}
-        ))
-      })
-
-      it('when a directory is renamed and a file is created in its place', async function () {
-        const reusedPath = path.join(watchDir, 'reused')
-        const newDirPath = path.join(watchDir, 'newdir')
-
-        await fs.mkdirs(reusedPath)
-        await until('directory creation event arrives', eventMatching(
-          {type: 'created', kind: 'directory', oldPath: reusedPath}
-        ))
-
-        await fs.rename(reusedPath, newDirPath)
-        await fs.writeFile(reusedPath, 'oh look a file\n')
-
-        await until('rename and creation events arrive', allEventsMatching(
-          {type: 'renamed', kind: 'directory', oldPath: reusedPath, newPath: newDirPath},
-          {type: 'created', kind: 'file', oldPath: reusedPath}
-        ))
-      })
-
-      it('when a directory is renamed and a file is renamed in its place', async function () {
-        const reusedPath = path.join(watchDir, 'reused')
-        const oldFilePath = path.join(watchDir, 'oldfile')
-        const newDirPath = path.join(watchDir, 'newdir')
-
-        await Promise.all([
-          fs.mkdirs(reusedPath),
-          fs.writeFile(oldFilePath, 'started as a file\n')
-        ])
-        await until('directory and file creation evenst arrive', allEventsMatching(
-          {type: 'created', kind: 'directory', oldPath: reusedPath},
-          {type: 'created', kind: 'file', oldPath: oldFilePath}
-        ))
-
-        await fs.rename(reusedPath, newDirPath)
-        await fs.rename(oldFilePath, reusedPath)
-
-        await until('rename events arrive', allEventsMatching(
-          {type: 'renamed', kind: 'directory', oldPath: reusedPath, newPath: newDirPath},
-          {type: 'renamed', kind: 'file', oldPath: oldFilePath, newPath: reusedPath}
-        ))
-      })
-
-      it('when a file is deleted and a directory is created in its place', async function () {
-        const reusedPath = path.join(watchDir, 'reused')
-        await fs.writeFile(reusedPath, 'something\n')
-        await until('directory creation event arrives', eventMatching(
-          {type: 'created', kind: 'file', oldPath: reusedPath}
-        ))
-
-        await fs.unlink(reusedPath)
-        await fs.mkdir(reusedPath)
-
-        await until('delete and create events arrive', orderedEventsMatching(
-          {type: 'deleted', oldPath: reusedPath},
-          {type: 'created', kind: 'directory', oldPath: reusedPath}
-        ))
-      })
-
-      it('when a file is deleted and a directory is renamed in its place', async function () {
-        const reusedPath = path.join(watchDir, 'reused')
-        const oldDirPath = path.join(watchDir, 'olddir')
-
-        await Promise.all([
-          fs.writeFile(reusedPath, 'something\n'),
-          fs.mkdir(oldDirPath)
-        ])
-        await until('creation events arrive', allEventsMatching(
-          {type: 'created', kind: 'file', oldPath: reusedPath},
-          {type: 'created', kind: 'directory', oldPath: oldDirPath}
-        ))
-
-        await fs.unlink(reusedPath)
-        await fs.rename(oldDirPath, reusedPath)
-
-        await until('delete and rename events arrive', allEventsMatching(
-          {type: 'deleted', oldPath: reusedPath},
-          {type: 'renamed', kind: 'directory', oldPath: oldDirPath, newPath: reusedPath}
-        ))
-      })
-
-      it('when a file is renamed and a directory is created in its place', async function () {
-        const reusedPath = path.join(watchDir, 'reused')
-        const newFilePath = path.join(watchDir, 'newfile')
-
-        await fs.writeFile(reusedPath, 'something\n')
-        await until('directory creation event arrives', eventMatching(
-          {type: 'created', kind: 'file', oldPath: reusedPath}
-        ))
-
-        await fs.rename(reusedPath, newFilePath)
-        await fs.mkdir(reusedPath)
-
-        await until('rename and create events arrive', orderedEventsMatching(
-          {type: 'renamed', kind: 'file', oldPath: reusedPath, newPath: newFilePath},
-          {type: 'created', kind: 'directory', oldPath: reusedPath}
-        ))
-      })
-
-      it('when a file is renamed and a directory is renamed in its place', async function () {
-        const reusedPath = path.join(watchDir, 'reused')
-        const oldDirPath = path.join(watchDir, 'olddir')
-        const newFilePath = path.join(watchDir, 'newfile')
-
-        await Promise.all([
-          fs.writeFile(reusedPath, 'something\n'),
-          fs.mkdir(oldDirPath)
-        ])
-        await until('file and directory creation events arrive', allEventsMatching(
-          {type: 'created', kind: 'file', oldPath: reusedPath},
-          {type: 'created', kind: 'directory', oldPath: oldDirPath}
-        ))
-
-        await fs.rename(reusedPath, newFilePath)
-        await fs.rename(oldDirPath, reusedPath)
-
-        await until('rename events arrive', orderedEventsMatching(
-          {type: 'renamed', kind: 'file', oldPath: reusedPath, newPath: newFilePath},
-          {type: 'renamed', kind: 'directory', oldPath: oldDirPath, newPath: reusedPath}
-        ))
-      })
-    })
-  })
-
   describe('unwatching a directory', function () {
     beforeEach(async function () {
       await watcher.configure({mainLog: mainLogFile, workerLog: workerLogFile})
